component table integration
