+++ src/app/shared/_component/table/log-table/log-table.component.scss
